requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Diffusion / ControlNet
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0

# Segmentation models
segmentation-models-pytorch>=0.3.3
timm>=0.9.0

# Image processing
opencv-python-headless>=4.8.0
albumentations>=1.3.1
scikit-image>=0.21.0
Pillow>=10.0.0
rembg>=2.0.50

# Background removal GPU support (optional)
# rembg[gpu]

# Scientific computing
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Geospatial
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0

# Memory-efficient optimizer
bitsandbytes>=0.41.0

# Progress bars & logging
tqdm>=4.65.0
wandb>=0.16.0

# Utilities
kaggle>=1.5.16
